+++ src/components/shop/OrdersView.vue
@@ -1,63 +1,14 @@
 <script setup lang="ts">
 import { ref, computed } from 'vue'
-import type { Order } from './components/OrderTypeDefine'
+import type { Order } from './components/orderTypeDefine'
 import OrderCard from './components/OrderCard.vue'
 import OrderDetailModal from './components/OrderDetailModal.vue'
+import { useOrders } from './components/useOrderModal'
+
+const { orders, selectedOrder, showModal, open, save } = useOrders()
 
 // 搜尋關鍵字
 const keyword = ref('')
-
-// 工單資料
-const orders = ref<Order[]>([
-  {
-    id: 'W-2024-001',
-    status: '待確認',
-    customer: { name: '王小明', phone: '0912-345-678' },
-    vehicle: { model: 'Toyota Camry 2020', service: '定期保養' },
-    date: '2024-12-18',
-    time: '09:30',
-    requestTime: '2024-12-10 14:20',
-  },
-  {
-    id: 'W-2024-002',
-    status: '進行中',
-    customer: { name: '李大華', phone: '0923-456-789' },
-    vehicle: { model: 'Honda CR-V 2019', service: '引擎檢修' },
-    date: '2024-12-18',
-    time: '10:00',
-    requestTime: '2024-12-11 09:50',
-  },
-  {
-    id: 'W-2024-003',
-    status: '已完成',
-    customer: { name: '張三', phone: '0933-222-333' },
-    vehicle: { model: 'BMW X5 2021', service: '換剎車皮' },
-    date: '2024-12-17',
-    time: '15:00',
-    requestTime: '2024-12-12 11:30',
-  },
-])
-
-// 浮窗控制
-const selectedOrder = ref<Order | null>(null)
-const showModal = ref(false)
-
-// 開啟浮窗
-const open = (order: Order) => {
-  selectedOrder.value = order
-  showModal.value = true
-}
-
-// 儲存修改
-const save = (status: Order['status']) => {
-  if (!selectedOrder.value) return
-
-  orders.value = orders.value.map((order) =>
-    order.id === selectedOrder.value!.id ? { ...order, status } : order,
-  )
-
-  showModal.value = false
-}
 
 // 搜尋過濾
 const filteredOrders = computed(() => {
