+++ src/components/shop/OverView.vue
@@ -1,60 +1,11 @@
 <script setup lang="ts">
-import { ref, computed } from 'vue'
-import type { Order } from './components/OrderTypeDefine'
+import { computed } from 'vue'
+import type { Order } from './components/orderTypeDefine'
+import { useOrders } from './components/useOrderModal'
 import OrderCard from './components/OrderCard.vue'
 import OrderDetailModal from './components/OrderDetailModal.vue'
 
-// 工單資料，寫死的假資料
-const orders = ref<Order[]>([
-  {
-    id: 'W-2026-001',
-    status: '進行中',
-    customer: {
-      name: '王小明',
-      phone: '0912-345-678',
-    },
-    vehicle: { model: 'Toyota Camry', service: '定期保養' },
-    requestTime: '2026-01-07 14:20',
-    date: '2026-01-09',
-    time: '09:30',
-  },
-  {
-    id: 'W-2026-002',
-    status: '待確認',
-    customer: {
-      name: '李大華',
-      phone: '0923-456-789',
-    },
-    vehicle: { model: 'Honda CR-V', service: '引擎檢修' },
-    requestTime: '2026-01-06 09:50',
-    date: '2026-01-09',
-    time: '10:00',
-  },
-  {
-    id: 'W-2026-003',
-    status: '已完成',
-    customer: {
-      name: '張三',
-      phone: '0933-222-333',
-    },
-    vehicle: { model: 'BMW X5', service: '煞車檢修' },
-    requestTime: '2026-01-05 11:15',
-    date: '2026-01-08',
-    time: '13:00',
-  },
-  {
-    id: 'W-2026-004',
-    status: '已完成',
-    customer: {
-      name: '林小七',
-      phone: '0988-777-666',
-    },
-    vehicle: { model: 'Audi A6', service: '輪胎更換' },
-    requestTime: '2026-01-02 10:00',
-    date: '2026-01-09',
-    time: '14:00',
-  },
-])
+const { orders, selectedOrder, showModal, open, save } = useOrders()
 
 const today = new Date()
 
@@ -92,25 +43,6 @@ const stats = computed(() => [
   { label: '待確認工單', value: waitingCount.value, icon: 'assignment' },
   { label: '本月完成', value: monthCompletedCount.value, icon: 'trending_up' },
 ])
-
-// 浮窗
-const showModal = ref(false)
-const selectedOrder = ref<Order | null>(null)
-
-const open = (order: Order) => {
-  selectedOrder.value = order
-  showModal.value = true
-}
-
-const save = (status: Order['status']) => {
-  if (!selectedOrder.value) return
-
-  orders.value = orders.value.map((order) =>
-    order.id === selectedOrder.value!.id ? { ...order, status } : order,
-  )
-
-  showModal.value = false
-}
 </script>
 
 <template>
